GPM module can be created from extracted metadata
The GPM module can be created from extracted metadata that is loaded
as a separate json file. This is helpful in cases where pdal strips
out the GPM metadata. Switched the encoding of numpy arrays in the
extracted GPM metadata to base64 to help support this. Also moved
the encoding methods to another file to clean things up.